requirements.txt
annotated-types==0.7.0
anyio==4.10.0
asgiref==3.9.1
Authlib==1.6.3
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
crispy-bootstrap5==2025.6
cryptography==45.0.7
deprecation==2.1.0
distro==1.9.0
Django==5.2.5
django-crispy-forms==2.4
greenlet==3.2.4
grpcio==1.74.0
h11==0.16.0
httpcore==1.0.9
httpx==0.27.0
idna==3.10
jiter==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.27
langchain-core==0.3.75
langchain-openai==0.3.32
langchain-text-splitters==0.3.9
langsmith==0.4.19
openai==1.102.0
orjson==3.11.3
packaging==25.0
pdf2image==1.17.0
pillow==11.3.0
protobuf==6.32.0
psycopg2-binary==2.9.10
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
PyPDF2==3.0.1
pytesseract==0.3.13
python-dotenv==1.1.1
PyYAML==6.0.2
regex==2025.7.34
requests==2.32.5
requests-toolbelt==1.0.0
sniffio==1.3.1
SQLAlchemy==2.0.43
sqlparse==0.5.3
tenacity==9.1.2
tiktoken==0.11.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
validators==0.28.1
weaviate-client==4.6.3
zstandard==0.24.0
sentence-transformers==3.0.1
scikit-learn
langchain-community==0.3.27
ollama
